requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
sentence-transformers>=2.6.0
numpy>=1.26.0
scipy>=1.12.0
openai>=1.20.0
httpx>=0.27.0
python-multipart>=0.0.9
openenv-core>=0.2.0
